QRWithoutQ

Documentation for QRWithoutQ.

QRWithoutQ._downcale_beta

QRWithoutQ._upscale_input!

QRWithoutQ.apply_reflector!

QRWithoutQ.clarfg!

QRWithoutQ.qr_unblocked!

QRWithoutQ.qrwoq!

QRWithoutQ.qrwoq_unblocked!

QRWithoutQ.safemin

QRWithoutQ._downcale_beta — Method

_downcale_beta(β, knt, sfmin)

Undo the rescaling of β.

QRWithoutQ._upscale_input! — Method

_upscale_input!(xtail::AbstractMatrix{T}, α, β, sfmin, rsfmin) where {T}

Rescale the input to clarfg! if it is too small.

QRWithoutQ.apply_reflector! — Method

apply_reflector!(x::AbstractVector, τ, A::AbstractVecOrMat)

QRWithoutQ.clarfg! — Method

clarfg!(x::AbstractVector{T}) where {T}

Intended to be a reasonably faithful implementation of clarg from LAPACK.

QRWithoutQ.qr_unblocked! — Method

qr_unblocked!(A::AbstractMatrix{T}, τ) where {T}

Computes the unblocked QR decomposition of A, in-place. The vector τ is used for storing the scalar coefficients for the Householder reflectors.

QRWithoutQ.qrwoq! — Method

qrwoq!(A::AbstractMatrix{T}, A::AbstractMatrix{T}) where {T}

Computes the upper triangular qr factor of A in-place and returns a view into A giving the upper triangular factor. The scalar factor the in

QRWithoutQ.qrwoq_unblocked! — Method

qrwoq_unblocked!(A::AbstractMatrix{T}) where {T}

Computes the R factor in the unblocked QR decomposition, in-place. R is returned as a view of A and all Householder vectors are destroyed.

QRWithoutQ.safemin — Method

safemin(::Type{T})

Smallest element of type T for which it is safe to divide by. Intended to be an implementation of LAPACKs slamch( 'S' ) / slamch( 'E' ) but I could be wrong...